requirements.txt
# conda create imageTemplateBot
# conda activate imageTemplateBot

python-telegram-bot
Pillow